Isolate logic on SparkPlanAnalyzer
